+++ tests/e2e/e2e_bypassminfee_test.go
@@ -4,6 +4,9 @@ import (
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	distributiontypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
+	ibcclienttypes "github.com/cosmos/ibc-go/v4/modules/core/02-client/types"
+	ibcchanneltypes "github.com/cosmos/ibc-go/v4/modules/core/04-channel/types"
+	"time"
 )
 
 func (s *IntegrationTestSuite) testBypassMinFeeWithdrawReward(endpoint string) {
@@ -94,4 +97,76 @@ func (s *IntegrationTestSuite) testBypassMinFeeWithdrawReward(endpoint string) {
 			s.Require().True(sdk.NewDecCoinFromCoin(incrBalance).IsGTE(oldBalancePlusReward[0]))
 		}
 	}
+}
+
+func (s *IntegrationTestSuite) testIBCBypassMsg() {
+
+	// submit gov proposal to change bypass-msg param to
+	// ["/ibc.core.channel.v1.MsgRecvPacket",
+	//  "/ibc.core.channel.v1.MsgAcknowledgement",
+	//  "/ibc.core.client.v1.MsgUpdateClient"]
+	submitterAddr := s.chainA.validators[0].keyInfo.GetAddress()
+	submitter := submitterAddr.String()
+	proposalCounter++
+	s.govProposeNewBypassMsgs([]string{
+		sdk.MsgTypeURL(&ibcchanneltypes.MsgRecvPacket{}),
+		sdk.MsgTypeURL(&ibcchanneltypes.MsgAcknowledgement{}),
+		sdk.MsgTypeURL(&ibcclienttypes.MsgUpdateClient{})}, proposalCounter, submitter, standardFees.String())
+
+	// use hermes1 to test default ibc bypass-msg
+	//
+	// test 1: transaction only contains bypass-msgs, pass
+	s.T().Logf("testing transaction contains only ibc bypass messages")
+	ok := s.hermesTransfer(hermesCofigWithGasPrices, s.chainA.id, s.chainB.id, transferChannel, uatomDenom, 100, 1000, 1)
+	s.Require().True(ok)
+
+	scrRelayerBalanceBefore, dstRelayerBalanceBefore := s.queryRelayerWalletsBalances()
+
+	pass := s.hermesClearPacket(hermesCofigNoGasPrices, s.chainA.id, transferChannel)
+	s.Require().True(pass)
+	pendingPacketsExist := s.hermesPendingPackets(hermesCofigNoGasPrices, s.chainA.id, transferChannel)
+	s.Require().False(pendingPacketsExist)
+
+	// confirm relayer wallets do not pay fees
+	scrRelayerBalanceAfter, dstRelayerBalanceAfter := s.queryRelayerWalletsBalances()
+	s.Require().Equal(scrRelayerBalanceBefore.String(), scrRelayerBalanceAfter.String())
+	s.Require().Equal(dstRelayerBalanceBefore.String(), dstRelayerBalanceAfter.String())
+
+	// test 2: test transactions contains both bypass and non-bypass msgs (sdk.MsgTypeURL(&ibcchanneltypes.MsgTimeout{})
+	s.T().Logf("testing transaction contains both bypass and non-bypass messages")
+	// hermesTransfer wtih --timeout-height-offset=1
+	ok = s.hermesTransfer(hermesCofigWithGasPrices, s.chainA.id, s.chainB.id, transferChannel, uatomDenom, 100, 1, 1)
+	s.Require().True(ok)
+	// make sure that the transaction is timeout
+	time.Sleep(3)
+	pendingPacketsExist = s.hermesPendingPackets(hermesCofigNoGasPrices, s.chainA.id, transferChannel)
+	s.Require().True(pendingPacketsExist)
+
+	pass = s.hermesClearPacket(hermesCofigNoGasPrices, s.chainA.id, transferChannel)
+	s.Require().False(pass)
+	// clear packets with paying fee, to not influence the next transaction
+	pass = s.hermesClearPacket(hermesCofigWithGasPrices, s.chainA.id, transferChannel)
+	s.Require().True(pass)
+
+	// test 3: test bypass-msgs exceed the MaxBypassGasUsage
+	s.T().Logf("testing bypass messages exceed MaxBypassGasUsage")
+	ok = s.hermesTransfer(hermesCofigWithGasPrices, s.chainA.id, s.chainB.id, transferChannel, uatomDenom, 100, 1000, 12)
+	s.Require().True(ok)
+	pass = s.hermesClearPacket(hermesCofigNoGasPrices, s.chainA.id, transferChannel)
+	s.Require().False(pass)
+
+	pendingPacketsExist = s.hermesPendingPackets(hermesCofigNoGasPrices, s.chainA.id, transferChannel)
+	s.Require().True(pendingPacketsExist)
+
+	pass = s.hermesClearPacket(hermesCofigWithGasPrices, s.chainA.id, transferChannel)
+	s.Require().True(pass)
+
+	// set the default bypass-msg back
+	proposalCounter++
+	s.govProposeNewBypassMsgs([]string{
+		sdk.MsgTypeURL(&ibcchanneltypes.MsgRecvPacket{}),
+		sdk.MsgTypeURL(&ibcchanneltypes.MsgAcknowledgement{}),
+		sdk.MsgTypeURL(&ibcclienttypes.MsgUpdateClient{}),
+		sdk.MsgTypeURL(&ibcchanneltypes.MsgTimeout{}),
+		sdk.MsgTypeURL(&ibcchanneltypes.MsgTimeoutOnClose{})}, proposalCounter, submitter, standardFees.String())
 }
